requirements.txt
flet>=0.7.0
flet_fastapi>=0.1.0
uvicorn>=0.23.0
cryptography>=39.0.0
zxcvbn>=4.4.28
requests>=2.28.0
pyobjc>=9.0.1;sys_platform=='darwin'
wmi>=1.5.1;sys_platform=='win32'
PyQt5>=5.15.0;sys_platform=='linux'
fprintd;sys_platform=='linux'
